Create cpu52.mac
This is the original 'intro-computer', a program made by Dec Education in Utrecht, Netherlands.
I once had the program printed out, and found the listing recently when cleaning the attic.
It took me a few evenings of re-entering the code from the listing 'as-is', i.e. including spelling errors etc.
This original version uses a VT100 in VT52 mode.